+++ chuffed/globals/graph.h
@@ -41,13 +41,13 @@ public:
 
 	inline BoolView& getNodeVar(int u) {
 		assert(u >= 0);
-		assert(u < vs.size());
+		assert(u < static_cast<int>(vs.size()));
 		return vs[u];
 	}
 
 	inline BoolView& getEdgeVar(int e) {
 		assert(e >= 0);
-		assert(e < es.size());
+		assert(e < static_cast<int>(es.size()));
 		return es[e];
 	}
 
@@ -57,26 +57,26 @@ public:
 
 	inline int getHead(int e) {
 		assert(e >= 0);
-		assert(e < es.size());
+		assert(e < static_cast<int>(es.size()));
 		return endnodes[e][1];
 	}
 
 	inline int getTail(int e) {
 		assert(e >= 0);
-		assert(e < es.size());
+		assert(e < static_cast<int>(es.size()));
 		return endnodes[e][0];
 	}
 
 	inline int getEndnode(int e, int n) {
 		assert(e >= 0);
-		assert(e < es.size());
+		assert(e < static_cast<int>(es.size()));
 		assert(n >= 0 && n < 2);
 		return endnodes[e][n];
 	}
 
 	inline int getOtherEndnode(int e, int u) {
 		assert(e >= 0);
-		assert(e < es.size());
+		assert(e < static_cast<int>(es.size()));
 		return endnodes[e][0] == u ? endnodes[e][1] : endnodes[e][0];
 	}
